Add lifecycle observer
