+++ coffee_maker_navigator_2/lib/app/app.dart
@@ -1,3 +1,4 @@
+import 'package:coffee_maker_navigator_2/app/app_lifecycle/ui/app_lifecycle_listenable.dart';
 import 'package:wolt_di/wolt_di.dart';
 import 'package:coffee_maker_navigator_2/app/di/coffee_maker_app_level_dependency_container.dart';
 import 'package:demo_ui_components/demo_ui_components.dart';
@@ -13,12 +14,14 @@ class CoffeeMakerApp extends StatelessWidget {
         final appLevelDependencyContainer = DependencyInjector.container<
             CoffeeMakerAppLevelDependencyContainer>(context);
 
-        return MaterialApp.router(
-          debugShowCheckedModeBanner: false,
-          theme: AppThemeData.themeData(context),
-          routerDelegate: appLevelDependencyContainer.appRouterDelegate,
-          backButtonDispatcher:
-              appLevelDependencyContainer.backButtonDispatcher,
+        return AppLifecycleObserver(
+          child: MaterialApp.router(
+            debugShowCheckedModeBanner: false,
+            theme: AppThemeData.themeData(context),
+            routerDelegate: appLevelDependencyContainer.appRouterDelegate,
+            backButtonDispatcher:
+                appLevelDependencyContainer.backButtonDispatcher,
+          ),
         );
       }),
     );
