+++ coffee_maker_navigator_2/lib/app/app_lifecycle/domain/app_lifecyle_service.dart
@@ -1,0 +1,25 @@
+import 'package:flutter/foundation.dart';
+
+enum AppLifeCycleState {
+  foreground,
+  background,
+}
+
+class AppLifeCycleService {
+  late ValueNotifier<AppLifeCycleState> lifeCycleState;
+
+  ValueListenable<AppLifeCycleState> get appLifeStateListenable =>
+      lifeCycleState;
+
+  AppLifeCycleService() {
+    lifeCycleState = ValueNotifier(AppLifeCycleState.foreground);
+  }
+
+  void onForegrounded() {
+    lifeCycleState.value = AppLifeCycleState.foreground;
+  }
+
+  void onBackgrounded() {
+    lifeCycleState.value = AppLifeCycleState.background;
+  }
+}
