+++ coffee_maker_navigator_2/lib/features/add_water/ui/view/widgets/add_water_screen_back_button.dart
@@ -1,0 +1,22 @@
+import 'package:flutter/material.dart';
+
+class AddWaterScreenBackButton extends StatelessWidget {
+  const AddWaterScreenBackButton({super.key});
+
+  @override
+  Widget build(BuildContext context) {
+    return Align(
+      alignment: Alignment.topLeft,
+      child: Padding(
+        padding: const EdgeInsets.all(16.0) +
+            EdgeInsets.only(top: MediaQuery.paddingOf(context).top),
+        child: BackButton(
+          style: ButtonStyle(
+            backgroundColor: MaterialStateProperty.all(Colors.white),
+          ),
+          onPressed: Navigator.of(context).pop,
+        ),
+      ),
+    );
+  }
+}
