+++ coffee_maker_navigator_2/lib/features/login/ui/view/login_screen.dart
@@ -1,6 +1,6 @@
 import 'package:coffee_maker_navigator_2/features/login/di/login_screen_dependency_container.dart';
+import 'package:coffee_maker_navigator_2/features/login/ui/view/widgets/login_screen_content.dart';
 import 'package:coffee_maker_navigator_2/features/login/ui/view_model/login_screen_view_model.dart';
-import 'package:demo_ui_components/demo_ui_components.dart';
 import 'package:flutter/material.dart';
 import 'package:wolt_di/wolt_di.dart';
 
@@ -13,57 +13,24 @@ class LoginScreen extends StatefulWidget {
 
 class _LoginScreenState extends State<LoginScreen>
     with
-        FeatureWithViewModelDependencyContainerSubscriptionMixin<
-            LoginScreenDependencyContainer, LoginScreenViewModel, LoginScreen> {
+        DependencyContainerSubscriptionMixin<LoginScreenDependencyContainer,
+            LoginScreen> {
+  late LoginScreenViewModel _viewModel;
+
+  @override
+  void initState() {
+    super.initState();
+    final container =
+        DependencyInjector.container<LoginScreenDependencyContainer>(context);
+    _viewModel = container.createViewModel();
+  }
+
   @override
   Widget build(BuildContext context) {
-    final textTheme = Theme.of(context).textTheme;
-
-    return Scaffold(
-      body: SafeArea(
-        child: Center(
-          child: SingleChildScrollView(
-            padding: const EdgeInsets.symmetric(horizontal: 40),
-            child: ConstrainedBox(
-              constraints: const BoxConstraints(maxWidth: 400),
-              child: Column(
-                mainAxisAlignment: MainAxisAlignment.center,
-                children: <Widget>[
-                  const Image(
-                    image: AssetImage('lib/assets/images/dash_coffee.webp'),
-                    fit: BoxFit.cover,
-                    height: 216,
-                    width: 384,
-                  ),
-                  Text(
-                    'Welcome to Coffee Maker!',
-                    style: textTheme.titleLarge!,
-                  ),
-                  const SizedBox(height: 50),
-                  const AppTextFormField(
-                    labelText: 'Username',
-                    textInputType: TextInputType.emailAddress,
-                  ),
-                  const SizedBox(height: 20),
-                  const AppTextFormField(
-                    labelText: 'Password',
-                    obscureText: true,
-                    autocorrect: false,
-                    textInputType: TextInputType.visiblePassword,
-                  ),
-                  const SizedBox(height: 30),
-                  WoltElevatedButton(
-                    onPressed: () {
-                      viewModel.onLoginPressed('email', 'password');
-                    },
-                    child: const Text('Sign in'),
-                  ),
-                ],
-              ),
-            ),
-          ),
-        ),
-      ),
+    return LoginScreenContent(
+      onLoginPressed: (email, password) {
+        _viewModel.onLoginPressed(email, password);
+      },
     );
   }
 }
