+++ coffee_maker_navigator_2/lib/features/login/ui/view_model/login_screen_view_model.dart
@@ -1,7 +1,6 @@
 import 'package:coffee_maker_navigator_2/app/auth/domain/auth_service.dart';
-import 'package:wolt_di/wolt_di.dart';
 
-class LoginScreenViewModel extends WoltViewModel {
+class LoginScreenViewModel {
   final AuthService authService;
 
   LoginScreenViewModel({required this.authService});
@@ -9,7 +8,4 @@ class LoginScreenViewModel extends WoltViewModel {
   Future<void> onLoginPressed(String email, String password) async {
     await authService.logIn(email, password);
   }
-
-  @override
-  void dispose() {}
 }
