+++ coffee_maker_navigator_2/lib/features/orders/di/orders_dependency_container.dart
@@ -1,11 +1,11 @@
+import 'package:coffee_maker_navigator_2/app/di/coffee_maker_app_level_dependency_container.dart';
 import 'package:coffee_maker_navigator_2/features/orders/data/remote/orders_remote_data_source.dart';
 import 'package:coffee_maker_navigator_2/features/orders/data/repository/orders_repository.dart';
 import 'package:coffee_maker_navigator_2/features/orders/domain/orders_service.dart';
 import 'package:coffee_maker_navigator_2/features/orders/ui/view_model/orders_screen_view_model.dart';
 import 'package:wolt_di/wolt_di.dart';
 
-class OrdersDependencyContainer
-    extends FeatureWithViewModelDependencyContainer {
+class OrdersDependencyContainer extends FeatureLevelDependencyContainer {
   late final OrdersRemoteDataSource _ordersRemoteDataSource;
   late final OrdersRepository _ordersRepository;
   late final OrdersService _ordersService;
@@ -19,10 +19,14 @@ class OrdersDependencyContainer
     _ordersRepository = OrdersRepository(
       ordersRemoteDataSource: _ordersRemoteDataSource,
     );
-    _ordersService = OrdersService(ordersRepository: _ordersRepository);
+    final appLevelDependencies =
+        bindWith<CoffeeMakerAppLevelDependencyContainer>();
+    _ordersService = OrdersService(
+      ordersRepository: _ordersRepository,
+      appLifeCycleService: appLevelDependencies.appLifeCycleService,
+    );
   }
 
-  @override
   OrdersScreenViewModel createViewModel() {
     return OrdersScreenViewModel(ordersService: _ordersService);
   }
