+++ wolt_di/lib/wolt_di.dart
@@ -3,7 +3,4 @@ export 'src/manager/dependency_container_manager.dart'
     hide DependencyContainerResolver, DependencyContainerRegistrar;
 export 'src/subscriber/dependency_container_subscription_mixin.dart';
 export 'src/subscriber/dependency_container_subscriber.dart';
-export 'src/view_model/feature_with_view_model_dependency_container.dart';
-export 'src/view_model/wolt_view_model.dart';
-export 'src/view_model/feature_with_view_model_dependency_container_subscription_mixin.dart';
 export 'src/dependency_injector.dart';
